add file for files and main file
