+++ tester_file/styles.css
@@ -3,6 +3,13 @@ body {
     background: green;
     width: 30px;
     height: 20px;
+    display:flex;
+    padding: 10rem;
+    margin: 5rem;
+    flex-wrap: wrap;
+    flex-direction: column;
+    background-color: yellow;
+    border: 10px;
 }
 
 .a {
